feat: Moves file input to public folder
